[Tweak] Restructured login and fuelday to match with the rest of the project - TKDATA-169
