+++ BilparkApp/src/components/fuelDay/FuelDayForm.js
@@ -3,7 +3,7 @@ import { connect } from 'react-redux';
 import { StyleSheet, Text, View } from 'react-native';
 import { Picker, CheckBox, ListItem, Body } from 'native-base';
 import { reduxForm, Field } from 'redux-form';
-import { postFuelDay } from '../actions/fuelDay';
+import { postFuelDay } from '../../actions/fuelDay';
 
 const renderPicker = ({ func, input, label, children, ...custom }) => (
   <Picker
